🌟 Added custom hook to debug recoil state
