+++ twake/frontend/src/app/features/debug/types/debug-recoil-state-types.ts
@@ -1,0 +1,5 @@
+export interface TwakeDebugState {
+  dumpStateSnapshot?(): void;
+  get?(key: string): void;
+  getAllAtoms?(): void;
+}
